requirements.txt
antlr4-python3-runtime>=4.13.0
networkx>=3.0
matplotlib>=3.5.0 